refactor: drop ark-node API (v1) support (#25)
