+++ lib/arkecosystem/client/connection.rb
@@ -1,18 +1,11 @@
 require 'arkecosystem/client/http/client'
-require 'arkecosystem/client/api/one/accounts'
-require 'arkecosystem/client/api/one/blocks'
-require 'arkecosystem/client/api/one/delegates'
-require 'arkecosystem/client/api/one/loader'
-require 'arkecosystem/client/api/one/peers'
-require 'arkecosystem/client/api/one/signatures'
-require 'arkecosystem/client/api/one/transactions'
-require 'arkecosystem/client/api/two/blocks'
-require 'arkecosystem/client/api/two/delegates'
-require 'arkecosystem/client/api/two/node'
-require 'arkecosystem/client/api/two/peers'
-require 'arkecosystem/client/api/two/transactions'
-require 'arkecosystem/client/api/two/votes'
-require 'arkecosystem/client/api/two/wallets'
+require 'arkecosystem/client/api/blocks'
+require 'arkecosystem/client/api/delegates'
+require 'arkecosystem/client/api/node'
+require 'arkecosystem/client/api/peers'
+require 'arkecosystem/client/api/transactions'
+require 'arkecosystem/client/api/votes'
+require 'arkecosystem/client/api/wallets'
 
 module ArkEcosystem
   module Client
@@ -39,27 +32,13 @@ module ArkEcosystem
         end
       end
 
-      # Return the Accounts API resource.
-      #
-      # @return [Object]
-      def accounts
-        case @version
-        when 1
-          @accounts ||= ArkEcosystem::Client::API::One::Accounts.new(@client)
-        else
-          raise NotImplementedError
-        end
-      end
-
       # Return the Blocks API resource.
       #
       # @return [Object]
       def blocks
         case @version
-        when 1
-          @blocks ||= ArkEcosystem::Client::API::One::Blocks.new(@client)
         when 2
-          @blocks ||= ArkEcosystem::Client::API::Two::Blocks.new(@client)
+          @blocks ||= ArkEcosystem::Client::API::Blocks.new(@client)
         else
           raise NotImplementedError
         end
@@ -70,22 +49,8 @@ module ArkEcosystem
       # @return [Object]
       def delegates
         case @version
-        when 1
-          @delegates ||= ArkEcosystem::Client::API::One::Delegates.new(@client)
         when 2
-          @delegates ||= ArkEcosystem::Client::API::Two::Delegates.new(@client)
-        else
-          raise NotImplementedError
-        end
-      end
-
-      # Return the Loader API resource.
-      #
-      # @return [Object]
-      def loader
-        case @version
-        when 1
-          @loader ||= ArkEcosystem::Client::API::One::Loader.new(@client)
+          @delegates ||= ArkEcosystem::Client::API::Delegates.new(@client)
         else
           raise NotImplementedError
         end
@@ -97,7 +62,7 @@ module ArkEcosystem
       def node
         case @version
         when 2
-          @node ||= ArkEcosystem::Client::API::Two::Node.new(@client)
+          @node ||= ArkEcosystem::Client::API::Node.new(@client)
         else
           raise NotImplementedError
         end
@@ -108,22 +73,8 @@ module ArkEcosystem
       # @return [Object]
       def peers
         case @version
-        when 1
-          @peers ||= ArkEcosystem::Client::API::One::Peers.new(@client)
         when 2
-          @peers ||= ArkEcosystem::Client::API::Two::Peers.new(@client)
-        else
-          raise NotImplementedError
-        end
-      end
-
-      # Return the Signatures API resource.
-      #
-      # @return [Object]
-      def signatures
-        case @version
-        when 1
-          @signatures ||= ArkEcosystem::Client::API::One::Signatures.new(@client) # rubocop:disable Metrics/LineLength
+          @peers ||= ArkEcosystem::Client::API::Peers.new(@client)
         else
           raise NotImplementedError
         end
@@ -134,22 +85,20 @@ module ArkEcosystem
       # @return [Object]
       def transactions
         case @version
-        when 1
-          @transactions ||= ArkEcosystem::Client::API::One::Transactions.new(@client) # rubocop:disable Metrics/LineLength
         when 2
-          @transactions ||= ArkEcosystem::Client::API::Two::Transactions.new(@client) # rubocop:disable Metrics/LineLength
+          @transactions ||= ArkEcosystem::Client::API::Transactions.new(@client) # rubocop:disable Metrics/LineLength
         else
           raise NotImplementedError
         end
       end
 
-      # Return the Vptes API resource.
+      # Return the Votes API resource.
       #
       # @return [Object]
       def votes
         case @version
         when 2
-          @votes ||= ArkEcosystem::Client::API::Two::Votes.new(@client)
+          @votes ||= ArkEcosystem::Client::API::Votes.new(@client)
         else
           raise NotImplementedError
         end
@@ -161,7 +110,7 @@ module ArkEcosystem
       def wallets
         case @version
         when 2
-          @wallets ||= ArkEcosystem::Client::API::Two::Wallets.new(@client)
+          @wallets ||= ArkEcosystem::Client::API::Wallets.new(@client)
         else
           raise NotImplementedError
         end
