+++ spec/client/api/blocks_spec.rb
@@ -1,4 +1,4 @@
-describe ArkEcosystem::Client::API::Two::Blocks do
+describe ArkEcosystem::Client::API::Blocks do
   before(:each) do
     get_v2_connection
   end
