+++ spec/client/api/node_spec.rb
@@ -1,4 +1,4 @@
-describe ArkEcosystem::Client::API::Two::Node do
+describe ArkEcosystem::Client::API::Node do
   before(:each) do
     get_v2_connection
   end
