+++ spec/client/api/peers_spec.rb
@@ -1,20 +1,20 @@
-describe ArkEcosystem::Client::API::Two::Votes do
+describe ArkEcosystem::Client::API::Peers do
   before(:each) do
     get_v2_connection
   end
 
   describe '.all' do
     it 'should be ok' do
-      response = @connection.votes.all
-      expect(response.url).to eq("#{@host}/votes")
+      response = @connection.peers.all
+      expect(response.url).to eq("#{@host}/peers")
       expect(response.body['success']).to be_truthy
     end
   end
 
   describe '.show' do
     it 'should be ok' do
-      response = @connection.votes.show('dummy')
-      expect(response.url).to eq("#{@host}/votes/dummy")
+      response = @connection.peers.show('dummy')
+      expect(response.url).to eq("#{@host}/peers/dummy")
       expect(response.body['success']).to be_truthy
     end
   end
