+++ spec/client/api/votes_spec.rb
@@ -1,20 +1,20 @@
-describe ArkEcosystem::Client::API::Two::Peers do
+describe ArkEcosystem::Client::API::Votes do
   before(:each) do
     get_v2_connection
   end
 
   describe '.all' do
     it 'should be ok' do
-      response = @connection.peers.all
-      expect(response.url).to eq("#{@host}/peers")
+      response = @connection.votes.all
+      expect(response.url).to eq("#{@host}/votes")
       expect(response.body['success']).to be_truthy
     end
   end
 
   describe '.show' do
     it 'should be ok' do
-      response = @connection.peers.show('dummy')
-      expect(response.url).to eq("#{@host}/peers/dummy")
+      response = @connection.votes.show('dummy')
+      expect(response.url).to eq("#{@host}/votes/dummy")
       expect(response.body['success']).to be_truthy
     end
   end
